// src/app/models/organization.ts
export class Organization {

    id: string;
    status: string;
    name: string;
    moxtraId: string;
    address: string;
    city: string;
    country: string;
    phoneNumber: string;
    numberOfActiveUsers: number;
    firstLogin: string;
    checkbox: boolean;
    ratePlan: string;
    expiryDate: string;
    timeZone: string;
    active: boolean;
    modeOfPayment: string;
    paymentStatus: string;
    numberOfCoreAgents: number;
    numberOfEnterpriseAgents: number;
    numberOfRecorderAgents: number;
    numberOfOutboundAgents: number;
    numberOfMonths: number;
}